Version: 2008-09-25 93t
Sourced from ast-open-history version 2008-09-25

+++ src/cmd/cs/fs_env.c
@@ -184,7 +184,7 @@ svc_read(void* handle, int fd)
 			err = ENOSYS;
 			break;
 		case MSG_read:
-			if (msg.argv[2].number < 0)
+			if ((long)msg.argv[2].number < 0)
 				err = EINVAL;
 			else if ((n = con->size - con->offset) > 0)
 			{
